Move tests/parser/* into the tests/ directory.
This was always kind of an arbitrary division.  I don't think it
actually did anything useful.
