+++ tests/pre-install.py
